requirements.txt
Flask==3.0.3
google-cloud-firestore==2.19.0
python-dotenv==1.0.1
gunicorn==22.0.0
